+++ packages/react-native-playground/.storybook/main.ts
@@ -9,7 +9,6 @@ const main: StorybookConfig = {
   addons: [
     "@storybook/addon-ondevice-notes",
     "@storybook/addon-ondevice-controls",
-    "@storybook/addon-ondevice-backgrounds",
     "@storybook/addon-ondevice-actions",
   ],
 };